Move lbPolicy execution profile tests to 'short', reduce to 4 hosts.
Also validate that each LBP is init'd once and adds a test to ensure
that the client loadBalancingPolicy is used when the default profile
does not have a loadBalancingPolicy.
